pyproject.toml
[project]
name = "ghostdrop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.69",
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
]
